read me file was changed, and a image was added to the images folder.
I worked on my part of the read me file, the team name and tag line are still undecided. Added a image to the images folder. Put a space for your part of the read me file. Added half of the ending of the read me file, from the example read me file on the course announcements.